reorganize pimms; AFQ attributes, BF
