+++ AFQ/definitions/mapping.py
@@ -7,7 +7,7 @@ from AFQ.definitions.utils import Definition, find_file
 from dipy.align import syn_registration, affine_registration
 import AFQ.registration as reg
 import AFQ.data as afd
-from AFQ.tasks.utils import as_file, get_fname
+from AFQ.tasks.utils import get_fname
 
 from dipy.align.imaffine import AffineMap
